// src/app/pages/vendors/dto/interfaces/i-vendor-detail.ts
export interface IVendorDetail {
  alipay: string;
  bankAccountNumber: string;
  bankInfo: string;
  brand: string;
  city: string;
  contactName: string;
  country: string;
  dateCreate: string;
  email: string;
  productTypes: string;
  purchasingVendorId: number;
  shippingAddress: string;
  telephone: string;
  vendorName: string;
  vendorStatus: number;
  wechatId: string;
  zip: string;
  hasUpdate: boolean;
}